Renamed `run_panos_upgrade` to `run_upgrade_device`
Renamed two major components in the code, `run_panos_upgrade` to `run_upgrade_device` and also the method `run_device_refresh` to `main`. The renamed functions are aimed at enhancing code readability and better reflect the operations they perform. There were also corresponding changes to the script imports and method call in tasks.py.
