+++ backend/panosupgradeweb/scripts/upgrade_device/upgrade.py
